+++ src/detection.hpp
@@ -2,7 +2,6 @@
 
 #include "opencv2/features2d/features2d.hpp"
 #include "opencv2/calib3d/calib3d.hpp"
-//#include "opencv2/nonfree/nonfree.hpp"
 #include "opencv2/xfeatures2d.hpp"
 #include "opencv2/core/core.hpp"
 #include "opencv2/imgproc/imgproc.hpp"
@@ -10,18 +9,18 @@
 #include "logger.hpp"
 
 
-#define MIN_POINTS 18 	// with less than that points object would not be found anyway
+#define MIN_POINTS 18    // with less than that points object would not be found anyway
 #define MIN_MATCHES 9   // less then this matcing points will cause false alarms
 
 
 #ifndef DETECTORS_H
 #define DETECTORS_H
 
-class Detector {
+class Detector
+{
 protected:
 	std::string object_id;
 	bool working = false;
-
 	Detector();
 
 public:
@@ -29,18 +28,17 @@ public:
 	bool isWorking() const;
 };
 
-class SiftDetector : public Detector {
+class SiftDetector : public Detector
+{
 	cv::Ptr<cv::Feature2D> features;
-
-	std::vector<cv::Point2f> obj_corners;
+	std::vector<cv::Point2d> obj_corners;
 	std::vector<cv::KeyPoint> keypoints;
 	cv::Mat descriptors;
 
 public:
 	SiftDetector(const std::string &object_id, int MinHess);
-
 	void process(cv::Mat image);
-	void match(const SiftDetector & sd_scene, const cv::Mat &img_scene) const;
+	void match(const SiftDetector &sd_scene, const cv::Mat &img_scene) const;
 
 //	static int SiftDetector::detections;
 };
